fixed some bugs with private data
