+++ dice_ml/explainer_interfaces/explainer_base.py
@@ -225,9 +225,6 @@ class ExplainerBase(ABC):
 
         self.check_query_instance_validity(features_to_vary, permitted_range, query_instance, feature_ranges_orig)
 
-        # check feature MAD validity and throw warnings
-        self.data_interface.check_mad_validity(feature_weights)
-
         return features_to_vary
 
     def check_query_instance_validity(self, features_to_vary, permitted_range, query_instance, feature_ranges_orig):
@@ -237,7 +234,6 @@ class ExplainerBase(ABC):
 
             if feature not in self.data_interface.feature_names:
                 raise ValueError("Feature", feature, "not present in training data!")
-
         for feature in self.data_interface.categorical_feature_names:
             if query_instance[feature].values[0] not in feature_ranges_orig[feature] and \
                     str(query_instance[feature].values[0]) not in feature_ranges_orig[feature]:
